requirements.txt
# LangChain 聊天系统依赖

# 核心框架
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-milvus>=0.1.0
langgraph>=0.0.30

# 数据库
pymilvus>=2.3.0
sqlite3

# 文本嵌入和模型
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 其他工具
python-dotenv>=1.0.0
requests>=2.28.0
typing-extensions>=4.5.0